+++ src/queries/index.ts
@@ -46,28 +46,28 @@ export function getThisMonthSavingsQuery(currency: string, rounding: number): st
 }
 
 export function getBalanceSheetQuery(currency: string): string {
-	return `SELECT account, convert(sum(position), '${currency}') WHERE account ~ '^(Assets|Liabilities|Equity)' AND NOT close_date(account) GROUP BY account ORDER BY account`;
+	return `SELECT account, number(only('${currency}', convert(sum(position), '${currency}'))) as operating_currency, subst('^\\([ ,]*', '(', subst('[ ,]*\\)$', ')', subst(', *[-.0-9]+ ${currency}|[-.0-9]+ ${currency} *,?', '', str(convert(sum(position), '${currency}'))))) as other_currencies WHERE account ~ '^(Assets|Liabilities|Equity)' AND NOT close_date(account) GROUP BY account ORDER BY account`;
 }
 
-export function getBalanceSheetQueryByCost(): string {
-	return `SELECT account, cost(sum(position)) WHERE account ~ '^(Assets|Liabilities|Equity)' AND NOT close_date(account) GROUP BY account ORDER BY account`;
+export function getBalanceSheetQueryByCost(currency: string): string {
+	return `SELECT account, number(only('${currency}', cost(sum(position)))) as operating_currency, subst('^\\([ ,]*', '(', subst('[ ,]*\\)$', ')', subst(', *[-.0-9]+ ${currency}|[-.0-9]+ ${currency} *,?', '', str(cost(sum(position)))))) as other_currencies WHERE account ~ '^(Assets|Liabilities|Equity)' AND NOT close_date(account) GROUP BY account ORDER BY account`;
 }
 
-export function getBalanceSheetQueryByUnits(): string {
-	return `SELECT account, units(sum(position)) WHERE account ~ '^(Assets|Liabilities|Equity)' AND NOT close_date(account) GROUP BY account ORDER BY account`;
+export function getBalanceSheetQueryByUnits(currency: string): string {
+	return `SELECT account, number(only('${currency}', units(sum(position)))) as operating_currency, subst('^\\([ ,]*', '(', subst('[ ,]*\\)$', ')', subst(', *[-.0-9]+ ${currency}|[-.0-9]+ ${currency} *,?', '', str(units(sum(position)))))) as other_currencies WHERE account ~ '^(Assets|Liabilities|Equity)' AND NOT close_date(account) GROUP BY account ORDER BY account`;
 }
 
 
 export function getIncomeStatementQuery(currency: string): string {
-	return `SELECT account, convert(sum(position), '${currency}') WHERE account ~ '^(Income|Expenses)' AND NOT close_date(account) GROUP BY account ORDER BY account`;
+	return `SELECT account, number(only('${currency}', convert(sum(position), '${currency}'))) as operating_currency, subst('^\\([ ,]*', '(', subst('[ ,]*\\)$', ')', subst(', *[-.0-9]+ ${currency}|[-.0-9]+ ${currency} *,?', '', str(convert(sum(position), '${currency}'))))) as other_currencies WHERE account ~ '^(Income|Expenses)' AND NOT close_date(account) GROUP BY account ORDER BY account`;
 }
 
-export function getIncomeStatementQueryByCost(): string {
-	return `SELECT account, cost(sum(position)) WHERE account ~ '^(Income|Expenses)' AND NOT close_date(account) GROUP BY account ORDER BY account`;
+export function getIncomeStatementQueryByCost(currency: string): string {
+	return `SELECT account, number(only('${currency}', cost(sum(position)))) as operating_currency, subst('^\\([ ,]*', '(', subst('[ ,]*\\)$', ')', subst(', *[-.0-9]+ ${currency}|[-.0-9]+ ${currency} *,?', '', str(cost(sum(position)))))) as other_currencies WHERE account ~ '^(Income|Expenses)' AND NOT close_date(account) GROUP BY account ORDER BY account`;
 }
 
-export function getIncomeStatementQueryByUnits(): string {
-	return `SELECT account, units(sum(position)) WHERE account ~ '^(Income|Expenses)' AND NOT close_date(account) GROUP BY account ORDER BY account`;
+export function getIncomeStatementQueryByUnits(currency: string): string {
+	return `SELECT account, number(only('${currency}', units(sum(position)))) as operating_currency, subst('^\\([ ,]*', '(', subst('[ ,]*\\)$', ')', subst(', *[-.0-9]+ ${currency}|[-.0-9]+ ${currency} *,?', '', str(units(sum(position)))))) as other_currencies WHERE account ~ '^(Income|Expenses)' AND NOT close_date(account) GROUP BY account ORDER BY account`;
 }
 
 export function getTransactionsQuery(filters: TransactionFilters, limit = 1000): string {
@@ -113,31 +113,31 @@ export function getBeanCheckCommand(filePath: string, commandBase: string): stri
 
 export function getHistoricalNetWorthDataQuery(interval: 'month' | 'week' = 'month', currency: string): string {
 	if (interval === 'week') {
-		return `SELECT last(date_add(date_trunc('week', date), 6)) AS week_end, only('${currency}', convert(last(balance), '${currency}', last(date_add(date_trunc('week', date), 6)))) WHERE account ~ '^(Assets|Liabilities)' GROUP BY date_trunc('week', date) ORDER BY week_end`;
+		return `SELECT last(date_add(date_trunc('week', date), 6)) AS week_end, number(only('${currency}', convert(last(balance), '${currency}', last(date_add(date_trunc('week', date), 6))))) WHERE account ~ '^(Assets|Liabilities)' GROUP BY date_trunc('week', date) ORDER BY week_end`;
 	}
-	return `SELECT year, month, only('${currency}', convert(last(balance), '${currency}', last(date_add(date(year + int(month/12), (month%12+1), 1), -1)))) WHERE account ~ '^(Assets|Liabilities)' GROUP BY year, month ORDER BY year, month`;
+	return `SELECT year, month, number(only('${currency}', convert(last(balance), '${currency}', last(date_add(date(year + int(month/12), (month%12+1), 1), -1))))) WHERE account ~ '^(Assets|Liabilities)' GROUP BY year, month ORDER BY year, month`;
 }
 
 
 export function getHistoricalNetProfitDataQuery(interval: 'month' | 'week' = 'month', currency: string): string {
 	if (interval === 'week') {
-		return `SELECT last(date_add(date_trunc('week', date), 6)) AS week_end, only('${currency}', convert(sum(position), '${currency}', last(date_add(date_trunc('week', date), 6)))) WHERE account ~ '^(Income|Expenses)' GROUP BY date_trunc('week', date) ORDER BY week_end`;
+		return `SELECT last(date_add(date_trunc('week', date), 6)) AS week_end, number(only('${currency}', convert(sum(position), '${currency}', last(date_add(date_trunc('week', date), 6))))) WHERE account ~ '^(Income|Expenses)' GROUP BY date_trunc('week', date) ORDER BY week_end`;
 	}
-	return `SELECT year, month, only('${currency}', convert(sum(position), '${currency}', last(date_add(date(year + int(month/12), (month%12+1), 1), -1)))) AS _worth WHERE account ~ '^(Income|Expenses)' GROUP BY year, month ORDER BY year, month`;
+	return `SELECT year, month, number(only('${currency}', convert(sum(position), '${currency}', last(date_add(date(year + int(month/12), (month%12+1), 1), -1))))) AS _worth WHERE account ~ '^(Income|Expenses)' GROUP BY year, month ORDER BY year, month`;
 }
 
 export function getHistoricalExpenseDataQuery(interval: 'month' | 'week' = 'month', currency: string): string {
 	if (interval === 'week') {
-		return `SELECT last(date_add(date_trunc('week', date), 6)) AS week_end, only('${currency}', convert(sum(position), '${currency}', last(date_add(date_trunc('week', date), 6)))) WHERE account ~ '^(Expenses)' GROUP BY date_trunc('week', date) ORDER BY week_end`;
+		return `SELECT last(date_add(date_trunc('week', date), 6)) AS week_end, number(only('${currency}', convert(sum(position), '${currency}', last(date_add(date_trunc('week', date), 6))))) WHERE account ~ '^(Expenses)' GROUP BY date_trunc('week', date) ORDER BY week_end`;
 	}
-	return `SELECT year, month, only('${currency}', convert(sum(position), '${currency}', last(date_add(date(year + int(month/12), (month%12+1), 1), -1)))) AS _worth WHERE account ~ '^(Expenses)' GROUP BY year, month ORDER BY year, month`;
+	return `SELECT year, month, number(only('${currency}', convert(sum(position), '${currency}', last(date_add(date(year + int(month/12), (month%12+1), 1), -1))))) AS _worth WHERE account ~ '^(Expenses)' GROUP BY year, month ORDER BY year, month`;
 }
 
 export function getHistoricalIncomeDataQuery(interval: 'month' | 'week' = 'month', currency: string): string {
 	if (interval === 'week') {
-		return `SELECT last(date_add(date_trunc('week', date), 6)) AS week_end, only('${currency}', convert(sum(position), '${currency}', last(date_add(date_trunc('week', date), 6)))) WHERE account ~ '^(Income)' GROUP BY date_trunc('week', date) ORDER BY week_end`;
+		return `SELECT last(date_add(date_trunc('week', date), 6)) AS week_end, number(only('${currency}', convert(sum(position), '${currency}', last(date_add(date_trunc('week', date), 6))))) WHERE account ~ '^(Income)' GROUP BY date_trunc('week', date) ORDER BY week_end`;
 	}
-	return `SELECT year, month, only('${currency}', convert(sum(position), '${currency}', last(date_add(date(year + int(month/12), (month%12+1), 1), -1)))) AS _worth WHERE account ~ '^(Income)' GROUP BY year, month ORDER BY year, month`;
+	return `SELECT year, month, number(only('${currency}', convert(sum(position), '${currency}', last(date_add(date(year + int(month/12), (month%12+1), 1), -1))))) AS _worth WHERE account ~ '^(Income)' GROUP BY year, month ORDER BY year, month`;
 }
 // --- List Budget/Target Quries ---	
 export function getBudgetListQuery(): string {